requirements.txt
pymongo[srv]==4.2.0
currencyconverter==0.17 
pandas==1.3.5
pyyaml==6.0
requests==2.28.1